break out firebase protocols
