+++ SwiftTesting/MockServices/FunctionsService.swift
@@ -1,0 +1,48 @@
+//
+//  FunctionsService.swift
+//  MovieClub
+//
+//  Created by Marcus Lair on 1/22/25.
+//
+
+import Testing
+import Firebase
+import Foundation
+import FirebaseFunctions
+import FirebaseFirestore
+import FirebaseStorage
+import FirebaseAuth
+@testable import MovieClub
+
+protocol FunctionsService {
+    // MARK: - Users
+    func createUserWithEmail(email: String, password: String, name: String) async throws -> String
+    func createUserWithOAuth(_ email: String, signInProvider: String) async throws -> String
+    func updateUser(userId: String, email: String?, displayName: String?) async throws
+    func deleteUser(_ id: String) async throws
+    
+    // MARK: - Comments
+    func postComment(movieId: String, text: String) async throws -> String
+    func likeComment(commentId: String) async throws
+    func unlikeComment(commentId: String) async throws
+    func deleteComment(commentId: String) async throws
+    
+    // MARK: - Suggestions
+    func createMovieClubSuggestion(clubId: String, suggestion: String) async throws -> String
+    func deleteMovieClubSuggestion(suggestionId: String) async throws
+    
+    // MARK: - Memberships
+    func joinMovieClub(clubId: String, userId: String) async throws
+    func leaveMovieClub(clubId: String, userId: String) async throws
+    
+    // MARK: - Movie Clubs
+    func createMovieClub(name: String, description: String) async throws -> String
+    func updateMovieClub(clubId: String, name: String?, description: String?) async throws
+    
+    // MARK: - Movies
+    func handleMovieReaction(movieId: String, reaction: String) async throws
+    func rotateMovie(movieId: String) async throws
+    
+    // MARK: - Posters
+    func collectPoster(poster: CollectionItem) async throws
+}
